+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 
 
 ## Unreleased
+- [Added] Allow ellipse shape output in the build script (#31)
 
 ## 2.0.4 - 2018-03-22
 - [Added] Program-management icon